+++ src/backend/CreateGroupSchedule.js
