+++ src/hooks/useWindowUnloadEffect.js
