+++ src/navigation/NavigationStack.js
@@ -15,18 +15,18 @@ import firebase from 'firebase/compat/app';
 import 'firebase/compat/auth';
 import 'firebase/compat/firestore';
 
-import HomeScreen from './Home';
-import CreateGroupScreen from './CreateGroup';
-import JoinGroupScreen from './JoinGroup';
-import GroupInfoScreen from './GroupInfo';
-import DrawerContent from './DrawerContent';
-import AvailabilityScreen from './Availability';
-import ScheduleScreen from './Schedule';
-import MonitorScreen from './Monitor';
-import InfoScreen from './Info';
-import ShiftsScreen from './Shifts';
-import AboutScreen from './About';
-import AccountSettingsScreen from './AccountSettings';
+import HomeScreen from '../screens/Home';
+import CreateGroupScreen from '../screens/CreateGroup';
+import JoinGroupScreen from '../screens/JoinGroup';
+import GroupInfoScreen from '../screens/GroupInfo';
+import DrawerContent from '../screens/DrawerContent';
+import AvailabilityScreen from '../screens/Availability';
+import ScheduleScreen from '../screens/Schedule';
+import MonitorScreen from '../screens/Monitor';
+import InfoScreen from '../screens/Info';
+import ShiftsScreen from '../screens/Shifts';
+import AboutScreen from '../screens/About';
+import AccountSettingsScreen from '../screens/AccountSettings';
 import { setCurrentUser, reset } from '../redux/reducers/userSlice';
 import { useTheme } from '../context/ThemeProvider';
 import { ActionSheetModal } from '../components/ActionSheetModal';
