+++ src/screens/Info.ios.js
